barrier: Add a new Meta wrapper for pointer barriers
Currently, we have a few function wrappers in the shell for pointer
barriers. If we want to implement pressure on a barrier, we need to
connect to some form of signal. In that case, we need to make a more
sophisticated wrapper for a pointer barrier. Add one, and stick it
in mutter.

https://bugzilla.gnome.org/show_bug.cgi?id=677215