wayland: rework buffer handling one more time
commit 0165cb6974 changed mutter to release shm buffers back to the
compositor as early as possible.  In order to track whether or not
a buffer can be released, the commit introduced an "accessible" flag
on the buffer structure (to mean "still being used, don't release
yet")

The problem, though, is that a buffer can theoretically be used by more
than one surface at a time.  If we have multiple surfaces delaying the
release of the buffer, then we need to know when every surface is done
with the buffer before doing the release.  A single boolean is
inappropriate for tracking whether or not multiple surfaces are delaying
release of the buffer, a usage counter is needed instead.

Furthermore, the surface objects don't need to access a
MetaWaylandBuffer object that's been released early—the texture created
from the buffer is referenced separately.  Since we don't need the
buffer after release, we can use the buffer references on the surface
objects directly as our usage counter to know when to release.

This commit drops meta_wayland_buffer_{take,release}_control and the
associated "accessible" struct member, and instead relies on the ref
count of the MetaWaylandBuffer itself to know when to do the release.

https://bugzilla.gnome.org/show_bug.cgi?id=761613