wayland: examine pending->buffer when prepping toplevel commit
toplevel_surface_commit looks at surface->buffer to decide how to
proceed.  It should instead look at pending->buffer, since pending
represents the new changes from the client to be committed.

At this point, the change may seem a bit academic, since surface->buffer
reflects changes made to the surface by the client, but in the future
we're going to nullify surface->buffer when the compositor is finished
with it, so this commit is important prep work for when the invariant
will no longer hold.

https://bugzilla.gnome.org/show_bug.cgi?id=761613