window: always disarm sync watchdog when client responds
At the moment we only disarm the watchdog timer set up
for sync requests if we're in the middle of a resize operation.

It's possible that the resize operation finished prematurely by
the user letting go of the mouse before the client updated.  If
that happens, when the client finally updates we'll leave the
timer engaged until it fires leading to resizes for the next
second to not get processed, and the client to get black listed
from future sync requests.